refactor: moved more structs to make_dto
