+++ jarust_plugins/examples/echotest_start_with_est.rs
@@ -8,8 +8,8 @@ use jarust_interface::tgenerator::RandomTransactionGenerator;
 use jarust_plugins::echo_test::events::EchoTestEvent;
 use jarust_plugins::echo_test::events::PluginEvent;
 use jarust_plugins::echo_test::jahandle_ext::EchoTest;
-use jarust_plugins::echo_test::msg_options::EchoTestStartParams;
-use jarust_plugins::echo_test::msg_options::EchoTestStartParamsRequired;
+use jarust_plugins::echo_test::params::EchoTestStartParams;
+use jarust_plugins::echo_test::params::EchoTestStartParamsRequired;
 use std::path::Path;
 use std::time::Duration;
 use tracing_subscriber::EnvFilter;
