+++ jarust_plugins/src/echo_test/mod.rs
@@ -1,4 +1,4 @@
 pub mod events;
 pub mod handle;
 pub mod jahandle_ext;
-pub mod msg_options;
+pub mod params;
